+++ sqlite3/lib/src/implementation/utils.dart
@@ -1,6 +1,5 @@
 import '../constants.dart';
 import 'bindings.dart';
-import 'exception.dart';
 
 extension BigIntRangeCheck on BigInt {
   BigInt get checkRange {
@@ -25,16 +24,6 @@ int eTextRep(bool deterministic, bool directOnly, bool subtype) {
   }
 
   return flags;
-}
-
-extension HandleResult<T extends Object> on SqliteResult<T> {
-  T okOrThrowOutsideOfDatabase(RawSqliteBindings bindings) {
-    if (resultCode != SqlError.SQLITE_OK) {
-      throw createExceptionOutsideOfDatabase(bindings, resultCode);
-    }
-
-    return result!;
-  }
 }
 
 extension ReadDartValue on RawSqliteValue {